requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.6.2.post1
astunparse==1.6.3
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
boto3==1.35.74
botocore==1.35.74
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cohere==5.5.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
fastapi-cli==0.0.5
fastavro==1.9.7
filelock==3.16.1
flatbuffers==25.1.24
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imgviz==1.7.6
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.8.0
kiwisolver==1.4.8
labelme==5.7.0
langchain==0.3.25
langchain-community==0.3.9
langchain-core==0.3.63
langchain-text-splitters==0.3.8
langgraph==0.3.34
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.63
langsmith==0.1.147
lazy_loader==0.4
libclang==18.1.1
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.10.1
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
natsort==8.4.0
networkx==3.4.2
numpy==1.26.4
ollama==0.2.1
onnxruntime==1.21.0
openai==1.30.2
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.12
ormsgpack==1.9.1
osam==0.2.2
packaging==24.2
parameterized==0.9.0
passlib==1.7.4
pillow==11.1.0
propcache==0.2.1
protobuf==5.29.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.1
pypdf==5.6.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.16
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
s3transfer==0.10.4
safetensors==0.5.2
scikit-image==0.25.2
scipy==1.15.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tifffile==2025.2.18
tiktoken==0.7.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.2
typer==0.12.5
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.1.3
win32_setctime==1.2.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
